gravity/requirements.txt
sentence-transformers>=3.0.0
anthropic>=0.40.0
numpy>=1.24.0
rich>=13.0.0
python-dotenv>=1.0.0
